refactor(antigravity): move port detection into a dedicated file
Move listeningPorts, its lsof/proc resolvers, and ProcNetTCPListeningPortParser
out of AntigravityStatusProbe.swift into AntigravityStatusProbe+PortDetection.swift,
following the existing +Deadline/+LocalEndpoints extension-file layout. This keeps
the primary probe file and type under the configured SwiftLint file_length and
type_body_length limits. No behavior change.
